Add Ruby projects to $PATH, Ruby load path.
Should serve as a ghetto way to ensure that I can use Ruby projects as
if they were fully installed (wrt scripts and require()s) without having
to constantly reinstall a gem.

Also won't get in the way when I do want to test actual installation of
a project's gem.

May replace this approach later with a symlink/develop addition to
either gem itself, or one of the gem helpers like Jeweler.